Convolutional Neural Network Architectures
Convolutional Neural Network Architectures for CIFAR 10 Image Classification using Tensorflow. 
The goal of the project was to design architectures for classification accuracy, experimenting with parameters such as layers, filter depths, convolutional layers and modules etc. The Architectures are as follows: 

- CNN Architecture 1: A Deeper Variation of a Basic CNN
- CNN Architecture 2: CNN Variation Inspired by the AlexNet Architecture 
- CNN Architecture 3: Inspiration from GoogleNet Inception Modules
- CNN Architecture 1 + 2: A Deeper variation combining CNN1 and CNN2

NB: Code is in first iteration and will be cleaned thoroughly for production quality.